requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-adk
google-genai
pypdf
python-docx
pytesseract
pdf2image
Pillow
cartesia
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client